% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_civic_api}
\alias{get_civic_api}
\title{GET from Civic API}
\usage{
get_civic_api(url, verbose)
}
\arguments{
\item{url}{the URL to GET}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
  \item status_code the HTTP status code (200 for success)
  \item response if status_code == 200, the JSON object;
    otherwise, the raw text.
}
}
\description{
GETs data from a Civic API endpoint
}
\examples{
\dontrun{
response <- rcivicapi::get_civic_api(
"https://api.mysportsfeeds.com/v2.0/pull/nba/2018-playoff/games.json"
)}
}
