% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{get_civic_api}
\alias{get_civic_api}
\title{GET from Civic API}
\usage{
get_civic_api(url, verbose)
}
\arguments{
\item{url}{the URL to GET}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
  \item `status_code` integer: the HTTP status code (200 for success)
  \item `response`: if `status_code` == 200, the JSON object from the API;
    otherwise, the raw text response.
}
}
\description{
GETs data from a Civic API endpoint
}
\examples{
\dontrun{
url_base <- "http://service.civicpdx.org/transportation2019/v1/toad"
endpoint <- "busPassengerStops"
query <- "?limit=10000&lines=10,14&stops=3637,3641,3633&time_range=9,10&service_key=W"
url <- paste(url_base, endpoint, query, sep = "/")
print(url)
response <- rcivicapi::get_civic_api(url, verbose = TRUE)
if (response[["status_code"]] == 200) {
  View(response[["data"]][["results"]][["features"]])
} else {
  print(response)
}
}
}
